Move json/yaml printer tests to correct location
