+++ staging/src/k8s.io/cli-runtime/pkg/printers/BUILD
@@ -38,6 +38,7 @@ go_library(
 go_test(
     name = "go_default_test",
     srcs = [
+        "json_test.go",
         "sourcechecker_test.go",
         "tableprinter_test.go",
         "template_test.go",
@@ -49,6 +50,10 @@ go_test(
         "//staging/src/k8s.io/apimachinery/pkg/apis/meta/v1/unstructured:go_default_library",
         "//staging/src/k8s.io/apimachinery/pkg/runtime:go_default_library",
         "//staging/src/k8s.io/apimachinery/pkg/runtime/schema:go_default_library",
+        "//staging/src/k8s.io/apimachinery/pkg/util/diff:go_default_library",
+        "//staging/src/k8s.io/apimachinery/pkg/util/json:go_default_library",
+        "//staging/src/k8s.io/client-go/kubernetes/scheme:go_default_library",
+        "//vendor/sigs.k8s.io/yaml:go_default_library",
     ],
 )
